refactor(dedup): extract handleInboundMessage into @multiwa/engine-runtime (step B)
Move the 335-line inbound-message body out of the API engine-manager into a
shared, dependency-injected handleInboundMessage(message, profileId, deps) in
@multiwa/engine-runtime. The API method is now a thin deps-builder that wires
its per-process collaborators (logger, engine access, resolveGroupName,
realtime emit, app-bus MESSAGE.RECEIVED, notification pre-scoped to MESSAGE,
rule-engine processMessage, AUTO_READ env) and delegates. prisma is imported
directly in the package (same pattern as ack-status). Removed the now-unused
serializeWaMessageId / applyInboundMedia / IncomingMessage imports from the API.

This is a PURE move — proven by the characterization suite from step A staying
15/15 GREEN against the delegating API method (deps route back to the same
collaborators). Behaviour on the live prod-active API path is unchanged.

Next (step C) repoints the worker onMessage to the same shared handler,
eliminating the ~315-line worker copy too.

engine-runtime build + api/worker tsc + 15/15 characterization + 130 module tests green.
