+++ packages/engine-runtime/src/index.ts
@@ -17,3 +17,4 @@ export * from './auto-reject-call';
 export * from './wa-message-id';
 export * from './system-message';
 export * from './engine-type';
+export * from './inbound-message';
